+++ media/app/components/TaskDetails/TaskDetails.vue
@@ -1,9 +1,11 @@
 <template>
 <div class="details">
+    <header class="details__header">Edit task as raw text:</header>
     <textarea ref="detailsTaskTitle"
               rows="1"
               class="details__input"
-              @input="onTaskTitleChange">{{ inputValue }}</textarea>
+              :value="inputValue"
+              @input="onTaskTitleChange"></textarea>
 </div>
 </template>