+++ src/commands/focusTasksWebviewAndInput.ts
@@ -1,9 +1,13 @@
 import { commands } from 'vscode';
 import { focusWebviewFilterInput } from '../webview/webviewView';
 
-export async function focusTasksWebviewAndInput() {
+export async function focusTasksWebviewAndInput(args: { selectInputText?: boolean; fillInputValue?: string }) {
 	await commands.executeCommand('todomd.webviewTasks.focus');
 	await commands.executeCommand('todomd.webviewTasks.focus');
 	await commands.executeCommand('todomd.webviewTasks.focus');
-	focusWebviewFilterInput();
+
+	focusWebviewFilterInput({
+		selectInputText: args?.selectInputText,
+		fillInputValue: args?.fillInputValue,
+	});
 }